fix(core): Execution status randomly stuck in "created"
The reason is that the ExecutionController try to fetch current execution on repository.
But since it was async call, the ExecutionController can return an `Flowable.empty()` on this case that will stop the webui.
Also fix infinite consume in case the execution is terminated, Now return the last execution if terminated.

close #48

close #48