feat(jdbc*): add an index on logs.timestamp
Fixes #2376

An issue exists in Postgres when we request ERROR logs on a wide timestamp range, it's cardinality calculation leads to ignore the existing indexes. Adding an index only on the timestamp column is a last resort for queries not finding existing indexes useful.
The same index is added to the other databases even if we didn't have clues on the same issue for consistency.